+++ DotNetLicenses/Configuration.fs
@@ -8,29 +8,44 @@ open System.Collections.Generic
 open System.IO
 open System.Threading.Tasks
 open DotNetLicenses.CommandLine
-open JetBrains.Annotations
 open Tomlyn
+open Tomlyn.Model
 
-[<CLIMutable>]
 type Configuration =
     {
-        MetadataSources: string[]
-        [<CanBeNull>] MetadataOverrides: Override[]
-        [<CanBeNull>] LockFile: string
-        [<CanBeNull>] PackagedFiles: PackageSpec[]
+        MetadataSources: MetadataSource[]
+        MetadataOverrides: Override[] option
+        LockFilePath: string option
+        PackagedFiles: PackageSpec[] option
     }
 
     static member Empty = {
         MetadataSources = Array.empty
-        MetadataOverrides = null
-        LockFile = null
-        PackagedFiles = null
+        MetadataOverrides = None
+        LockFilePath = None
+        PackagedFiles = None
     }
 
     static member Read(stream: Stream, filePath: string option): Task<Configuration> = task {
+        let readSources table = failwith "TODO"
+        let readOverrides = failwith "TODO"
+        let readPackagedFiles = failwith "TODO"
+
         use reader = new StreamReader(stream)
         let! text = reader.ReadToEndAsync()
-        return Toml.ToModel<Configuration>(text, Option.toObj filePath, options = null)
+        let table = Toml.ToModel(text, sourcePath = Option.toObj filePath)
+
+        let sources = table["metadata_sources"] :?> TomlArray
+        let overrides = table["metadata_overrides"] :?> TomlArray
+        let lockFilePath = table["lock_file"] :?> string
+        let packagedFiles = table["packaged_files"] :?> TomlArray
+
+        return {
+            MetadataSources = readSources sources
+            MetadataOverrides = readOverrides overrides
+            LockFilePath = Option.ofObj lockFilePath
+            PackagedFiles = readPackagedFiles packagedFiles
+        }
     }
 
     static member ReadFromFile(path: string): Task<Configuration> = task {
@@ -51,7 +66,7 @@ type Configuration =
         }
 
         let map = Dictionary()
-        Option.ofObj this.MetadataOverrides
+        this.MetadataOverrides
         |> Option.defaultValue Array.empty
         |> Seq.map(fun o -> packageReference o, metadataOverride o)
         |> Seq.iter(fun (k, v) ->
@@ -64,13 +79,16 @@ type Configuration =
         )
         map
 
-and [<CLIMutable>] Override = {
+and MetadataSource =
+    NuGet of ``include``: string
+
+and Override = {
     Id: string
     Version: string
     Spdx: string
     Copyright: string
 }
-and [<CLIMutable>] PackageSpec = {
-    Type: string
-    Path: string
-}
+
+and PackageSpec =
+    | Directory of path: string
+    | Zip of path: string