+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add configuration file support, the file path being passed through the command line.
 
   Configuration file options available in this version: `metadata_sources`, `metadata_overrides`, `lock_file`, and `packaged_files`.
+- Supported metadata sources: `nuget`.
 - License lock file support for the packaged files.
 - Supported commands: `print` (default) and `generate-lock`.
 - Warnings and non-zero exit codes on duplicate and unused overrides.